Add header, main and footer tags
These tags provide essential landmarks on the page for a11y software to use.
