+++ static/global.css
@@ -22,7 +22,6 @@ html {
 .container {
 	display: flex;
 	flex-direction: column;
-	height: 100%;
 }
 
 .section {
